OrderTest & OrderDetaTest

+++ MvcOnlineStore/StoreDataLibary/Models/OrderDetails.cs
@@ -5,6 +5,6 @@
         public string OrderID { get; set; }
         public string ProductID { get; set; }
         public int Quantity { get; set; }
-        public int Discount { get; set; }
+        public double Discount { get; set; }
     }
 }
